rename packageInTransitScreen to packageInTransit
